requirements.txt
pip-requirements
gpt4all
pyttsx3
SpeechRecognition
customtkinter
Pillow
numpy
psutil
wikipedia

pyaudio
edge-tts
protobuf==3.20.3
pytest
python-dotenv
cryptography
matplotlib
openai-whisper
torch
